pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strategy-backtester"
version = "0.1.0"
description = "Strategy Backtesting Engine (Alpha Vantage, MA/Mom/MeanRev, metrics, MC, optimization)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pytest",
  "hydra-core",
  "mlflow",
  "optuna",
  "scikit-learn",
  "scipy",
  "numpy>=1.25",
  "pandas>=2.1",
  "matplotlib>=3.8",
  "plotly>=5.20",
  "requests>=2.32",
  "python-dotenv>=1.0",
  "tqdm>=4.66"
]

[project.scripts]
sbe = "strategy_backtester.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
